Refactor: Move command tests into a package
