+++ src/main/java/bobbybot/enums/BotCommand.java
@@ -4,5 +4,5 @@ package bobbybot.enums;
  * Enum for all valid Commands
  */
 public enum BotCommand {
-    BYE, LIST, DONE, DEADLINE, TODO, EVENT, DELETE, FIND, CONTACT, LIST_CONTACT
+    BYE, LIST, DONE, DEADLINE, TODO, EVENT, DELETE, FIND, CONTACT, LIST_CONTACT, DELETE_CONTACT
 }
